requirements.txt
requests
diskcache
python-dotenv